+++ apps/www/content/docs/upload.mdx
@@ -6,27 +6,22 @@ docs:
     title: Upload
 ---
 
+### UploadThing Integration
 
-<ComponentPreview name="playground-demo" id="upload" />
+Make sure you have install the [media-placeholder-element](/docs/components/media-placeholder-element) component and all the dependencies.
 
-{/* ### UploadThing Integration
+Set `UPLOADTHING_TOKEN` in your .env file [get one here](https://uploadthing.com/dashboard).
 
-This component uses UploadThing for file uploads. UploadThing provides a simple and efficient way to handle file uploads in your application.
 
-To use UploadThing:
+### Using your own backend
 
-1. Set up an UploadThing account and configure your upload endpoints.
-2. Install the UploadThing client in your project:
+Remove this two folder `lib/uploadthing` and `/api/uploadthing`.
 
-```bash
-npm install uploadthing
-```
-
-3. Configure the UploadThing client in your application.
-
-For more details on setting up UploadThing, refer to their [documentation](https://docs.uploadthing.com/). */}
+Then impelement a similar hooks like `useUploadFile` using your own backend.
 
 ## Examples
+
+<ComponentPreview name="playground-demo" id="mediaPlaceholder" />
 
 ### Plate UI
