+++ apps/www/src/lib/plate/demo/values/mediaValue.tsx
@@ -22,16 +22,40 @@ export const imageValue: any = (
 export const mediaPlaceholderValue: any = (
   <fragment>
     <hh2>Upload</hh2>
-    <hp indent={1} listStyleType="disc">
-      Easily upload media files by dragging and dropping them into the editor or
-      using the file picker. The editor provides:
+    <hp>
+      Our editor supports various media types for upload, including images,
+      videos, audio, and files.
     </hp>
+    <hfile
+      name="sample.pdf"
+      align="center"
+      url="https://s26.q4cdn.com/900411403/files/doc_downloads/test.pdf"
+      width="80%"
+      isUpload
+    >
+      <htext />
+    </hfile>
     <hp indent={1} listStyleType="disc">
       Real-time upload status and progress tracking
     </hp>
+    <haudio
+      align="center"
+      url="https://samplelib.com/lib/preview/mp3/sample-3s.mp3"
+      width="80%"
+    >
+      <htext />
+    </haudio>
     <hp indent={1} listStyleType="disc">
       Configurable file size limits and batch upload settings
     </hp>
+    <hvideo
+      align="center"
+      url="https://videos.pexels.com/video-files/6769791/6769791-uhd_2560_1440_24fps.mp4"
+      width="80%"
+      isUpload
+    >
+      <htext />
+    </hvideo>
     <hp indent={1} listStyleType="disc">
       Clear error messages for any upload issues
     </hp>
