+++ apps/www/src/registry/registry-ui.ts
@@ -845,6 +845,40 @@ export const uiNodes: Registry = [
     type: 'registry:ui',
   },
   {
+    dependencies: [
+      '@udecode/plate-media',
+      'use-file-picker',
+      '@uploadthing/react@7.1.0',
+      'uploadthing@7.2.0',
+      'zod',
+      'sonner',
+    ],
+    doc: {
+      description: 'A placeholder for media files.',
+      docs: [{ route: '/docs/media-placeholder', title: 'Media Placeholder' }],
+      examples: ['media-demo', 'media-toolbar-pro'],
+    },
+    files: [
+      '../../../../../templates/plate-playground-template/src/components/plate-ui/media-placeholder-element.tsx',
+      '../../../../../templates/plate-playground-template/src/lib/uploadthing/uploadthing.ts',
+    ],
+    name: 'media-placeholder-element',
+    registryDependencies: [],
+    type: 'registry:ui',
+  },
+  {
+    dependencies: [],
+    doc: {
+      description: 'A toast for media uploads.',
+      docs: [{ route: '/docs/media-placeholder', title: 'Media Placeholder' }],
+      examples: ['media-demo', 'upload-pro'],
+    },
+    files: ['plate-ui/media-upload-toast.tsx'],
+    name: 'media-upload-toast',
+    registryDependencies: [],
+    type: 'registry:ui',
+  },
+  {
     dependencies: [],
     doc: {
       description: 'A quote component for block quotes.',
