+++ templates/plate-playground-template/src/lib/uploadthing/index.ts
@@ -1,5 +1,1 @@
-export * from './handle-error';
-
 export * from './uploadthing';
-
-export * from './use-upload-file';
