+++ templates/plate-playground-template/src/lib/uploadthing/uploadthing.ts
@@ -1,6 +1,102 @@
+import * as React from 'react';
+
 import type { OurFileRouter } from '@/app/api/uploadthing/core';
+import type {
+  ClientUploadedFileData,
+  UploadFilesOptions,
+} from 'uploadthing/types';
 
 import { generateReactHelpers } from '@uploadthing/react';
+import { isRedirectError } from 'next/dist/client/components/redirect';
+import { toast } from 'sonner';
+import { z } from 'zod';
 
+export interface UploadedFile<T = unknown> extends ClientUploadedFileData<T> {}
+
+interface UseUploadFileProps
+  extends Pick<
+    UploadFilesOptions<OurFileRouter, keyof OurFileRouter>,
+    'headers' | 'onUploadBegin' | 'onUploadProgress' | 'skipPolling'
+  > {
+  onUploadComplete?: (file: UploadedFile) => void;
+  onUploadError?: (error: unknown) => void;
+}
+
+export function useUploadFile(
+  endpoint: keyof OurFileRouter,
+  { onUploadComplete, onUploadError, ...props }: UseUploadFileProps = {}
+) {
+  const [uploadedFile, setUploadedFile] = React.useState<UploadedFile>();
+  const [uploadingFile, setUploadingFile] = React.useState<File>();
+  const [progress, setProgress] = React.useState<number>(0);
+  const [isUploading, setIsUploading] = React.useState(false);
+
+  async function uploadThing(file: File) {
+    setIsUploading(true);
+    setUploadingFile(file);
+
+    try {
+      const res = await uploadFiles(endpoint, {
+        ...props,
+        files: [file],
+        onUploadProgress: ({ progress }) => {
+          setProgress(Math.min(progress, 100));
+        },
+      });
+
+      setUploadedFile(res[0]);
+
+      onUploadComplete?.(res[0]);
+
+      return uploadedFile;
+    } catch (error) {
+      const errorMessage = getErrorMessage(error);
+
+      const message =
+        errorMessage.length > 0
+          ? errorMessage
+          : 'Something went wrong, please try again later.';
+
+      toast.error(message);
+      onUploadError?.(error);
+    } finally {
+      setProgress(0);
+      setIsUploading(false);
+      setUploadingFile(undefined);
+    }
+  }
+
+  return {
+    isUploading,
+    progress,
+    uploadFile: uploadThing,
+    uploadedFile,
+    uploadingFile,
+  };
+}
 export const { uploadFiles, useUploadThing } =
   generateReactHelpers<OurFileRouter>();
+
+export function getErrorMessage(err: unknown) {
+  const unknownError = 'Something went wrong, please try again later.';
+
+  if (err instanceof z.ZodError) {
+    const errors = err.issues.map((issue) => {
+      return issue.message;
+    });
+
+    return errors.join('\n');
+  } else if (err instanceof Error) {
+    return err.message;
+  } else if (isRedirectError(err)) {
+    throw err;
+  } else {
+    return unknownError;
+  }
+}
+
+export function showErrorToast(err: unknown) {
+  const errorMessage = getErrorMessage(err);
+
+  return toast.error(errorMessage);
+}
